requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.1
aiofiles>=0.7.0
python-dotenv>=0.19.0
requests>=2.26.0
gradio>=4.0.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0